move RuleConfigurationsYamlSwapper
